Check these edits before merging! (#5)
* Check these edits before merging!

These are my _guesses_ about the data format and need to be verified by someone who knows better.

(That example body temperature, for instance, would not be a very healthy one. What do _typical_ values look like? Are absolute times, for ISO 8601, known or is it the number of seconds since the device was turned on? Or are the times reported as absolute but actually undetermined up to an additive constant?)

* Corrections from @henryiii.